Removed dependency on KnockoutJS
jsSort no longer requires KnockoutJS to be loaded. If a property that is to be sorted on is a funciton, the function will be called. This handles knockout observables but also handles any other function property.

In sample code I've removed the 3rd party libraries and instead load them from cdnjs. The sample still uses knockout but this is solely to demonstrate support for obserable properties still works and to build up the UI for the sample.

Udpated readme file to reflect removal of Knockout as a dependency.